removed lib/. refactored UI look&feel.

+++ src/ui/Screen.h
@@ -39,5 +39,5 @@ protected:
 
     virtual void onHide() {};
 private:
-    size_t firstDisplayedMenuItem;
+    uint8_t firstDisplayedMenuItem;
 };
